* commit '2553c49814aa7fe55bc73e782449c882b9fd4d45':
  挂模板、分享
  bug
  干了这碗热翔
  bug

+++ fbi_engine_analyse/src/com/fr/bi/field/dimension/dimension/BINumberDimension.java
@@ -56,7 +56,7 @@ public class BINumberDimension extends BIAbstractDimension {
     @Override
     public Object getValueByType(Object data) {
         if (group.getType() == BIReportConstant.GROUP.ID_GROUP){
-            return data == null ? null : Long.parseLong(data.toString());
+            return data == null ? null : Double.parseDouble(data.toString());
         }
         return data == null ? StringUtils.EMPTY : data.toString();
     }
